File structure + abort linefeed
