+++ Cargo.toml
@@ -14,14 +14,13 @@ license = "EUPL-1.2"
 exclude = ["/target", "/.vscode/", "/.github/", "/target/", "/SECURITY.md"]
 
 [dependencies]
-linefeed = { version = "0.6.0", optional = true } 
 reqwest = { version = "0.11.11", optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
 serde_json = { version = "1.0", optional = true }
 
 [features]
 default = ["printer", "orochi-repo"]
-printer = ["linefeed"]
+printer = []
 orochi = ["serde", "serde_json"]
 orochi-repo = ["orochi", "reqwest"]
