+++ src/orochi/repository/http.rs
@@ -1,0 +1,3 @@
+//! HTTP Repository implementation
+//! 
+//! This module implements utils for HTTP repositores.
