+++ src/orochi/repository/mod.rs
@@ -1,0 +1,14 @@
+/// FTP Repository implementation
+/// 
+/// This module implements utils for FTP repositores.
+mod ftp;
+
+/// HTTP Repository implementation
+/// 
+/// This module implements utils for HTTP repositores.
+mod http;
+
+/// SSH Repository implementation
+/// 
+/// This module implements utils for SSH repositores.
+mod ssh;
